# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Gateway for School Online service.",
        "title": "School Online API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/user": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Create user",
                "parameters": [
                    {
                        "description": "User data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_abozorov_school_online_cmd_api_gateway_internal_models.RegisterUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer",
                                "format": "int32"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update user",
                "parameters": [
                    {
                        "description": "Updated user data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_abozorov_school_online_cmd_api_gateway_internal_models.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/user/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "List users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_abozorov_school_online_cmd_api_gateway_internal_models.User"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/user/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_abozorov_school_online_cmd_api_gateway_internal_models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Delete user by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_abozorov_school_online_cmd_api_gateway_internal_models.ParentRole": {
            "type": "object",
            "properties": {
                "students_id": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "github_com_abozorov_school_online_cmd_api_gateway_internal_models.RegisterUserRequest": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent_role": {
                    "$ref": "#/definitions/github_com_abozorov_school_online_cmd_api_gateway_internal_models.ParentRole"
                },
                "password": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "staff_role": {
                    "$ref": "#/definitions/github_com_abozorov_school_online_cmd_api_gateway_internal_models.StaffRole"
                },
                "student_role": {
                    "$ref": "#/definitions/github_com_abozorov_school_online_cmd_api_gateway_internal_models.StudentRole"
                },
                "teacher_role": {
                    "$ref": "#/definitions/github_com_abozorov_school_online_cmd_api_gateway_internal_models.TeacherRole"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "github_com_abozorov_school_online_cmd_api_gateway_internal_models.StaffRole": {
            "type": "object",
            "properties": {
                "experience": {
                    "type": "integer"
                },
                "position": {
                    "type": "string"
                }
            }
        },
        "github_com_abozorov_school_online_cmd_api_gateway_internal_models.StudentRole": {
            "type": "object",
            "properties": {
                "classroom_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_abozorov_school_online_cmd_api_gateway_internal_models.TeacherRole": {
            "type": "object",
            "properties": {
                "experience": {
                    "type": "integer"
                },
                "subjects_id": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "github_com_abozorov_school_online_cmd_api_gateway_internal_models.UpdateUserRequest": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parent_role": {
                    "$ref": "#/definitions/github_com_abozorov_school_online_cmd_api_gateway_internal_models.ParentRole"
                },
                "phone_number": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "staff_role": {
                    "$ref": "#/definitions/github_com_abozorov_school_online_cmd_api_gateway_internal_models.StaffRole"
                },
                "student_role": {
                    "$ref": "#/definitions/github_com_abozorov_school_online_cmd_api_gateway_internal_models.StudentRole"
                },
                "teacher_role": {
                    "$ref": "#/definitions/github_com_abozorov_school_online_cmd_api_gateway_internal_models.TeacherRole"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_abozorov_school_online_cmd_api_gateway_internal_models.User": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parent_role": {
                    "$ref": "#/definitions/github_com_abozorov_school_online_cmd_api_gateway_internal_models.ParentRole"
                },
                "phone_number": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "staff_role": {
                    "$ref": "#/definitions/github_com_abozorov_school_online_cmd_api_gateway_internal_models.StaffRole"
                },
                "student_role": {
                    "$ref": "#/definitions/github_com_abozorov_school_online_cmd_api_gateway_internal_models.StudentRole"
                },
                "teacher_role": {
                    "$ref": "#/definitions/github_com_abozorov_school_online_cmd_api_gateway_internal_models.TeacherRole"
                },
                "username": {
                    "type": "string"
                },
                "verify_email": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by a space and",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}